requirements.txt
pandas 
textblob
plotly.
nltk
sklearn.feature_extraction.text
numpy
talib
textblob
